requirements.txt
-i https://pypi.org/simple
aiocontextvars==0.2.2 ; python_version < '3.7'
appnope==0.1.0 ; sys_platform == 'darwin'
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.0
contextvars==2.4 ; python_version < '3.7'
decorator==4.4.1
escapism==1.0.0
flask==1.1.1
google-auth==1.11.0
gunicorn==20.0.4
idna==2.8
immutables==0.11
ipython-genutils==0.2.0
ipython==7.11.1
itsdangerous==1.1.0
jedi==0.16.0
jinja2==2.11.0
jsonpatch==1.25
jsonpointer==2.0
kubernetes==9.0.1
loguru==0.4.1
markupsafe==1.1.1
oauthlib==3.1.0
parso==0.6.0
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pygments==2.5.2
python-dateutil==2.8.1
pyyaml==5.3
requests-oauthlib==1.3.0
requests==2.22.0
rsa==4.0
six==1.14.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
websocket-client==0.57.0
werkzeug==0.16.1
